server/requirements.txt
flask==2.2.3
Pillow==9.5.0
